% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/templates.R
\name{Simulation_Skeleton}
\alias{Simulation_Skeleton}
\title{Open a simulation skeleton}
\usage{
Simulation_Skeleton(filename)
}
\arguments{
\item{filename}{Name of file to create.}
}
\description{
Creates and opens an .R file containing a skeleton for writing a Monte Carlo simulation study.
}
\examples{
\dontrun{
Simulation_Skeleton("Behrens-Fisher problem")
}
}

