% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel-setup.R
\name{start_parallel}
\alias{start_parallel}
\title{Initializes a cluster for parallel computing}
\usage{
start_parallel(source_obj = NULL, cores)
}
\arguments{
\item{source_obj}{list of objects to be exported to each node of the cluster. Defaults to NULL.}

\item{cores}{desired number of cores. Defaults to one less than the number of available cores.}
}
\value{
On Windows, returns the cluster object. On Mac, returns NULL.
}
\description{
On Windows, this function initializes a socket cluster, registers the cluster with the DoParallel package,
exports a list of objects to each node of the cluster, and returns the cluster object. On Mac
and other operating systems that can take advantage of multicore, the function simply registers a cluster
of the specified number of nodes with DoParallel.
}
\examples{
\dontrun{
cluster <- start_parallel()
stopCluster(cluster)
}
}

